+++ test/test087.c
@@ -12,14 +12,16 @@
 // *TW
 //===---------------------------------------------------------------------------===//
 
+#include <stdint.h>
+
 int main() {
 
   char inc0 = 0, inc1 = 0;
   int diff = 0, a = 100;
   int *p = &a;
-  inc0 =(int)p;
+  inc0 =(intptr_t)p;
   --(*p--);  //--*p--);
-  inc1 =(int)p;
+  inc1 =(intptr_t)p;
   diff =  inc0-inc1;
   diff += 2;
   return diff;
